Gemini-Bot-backend/requirements.txt
Django
djangorestframework
google-genai
python-decouple
whitenoise
